refactor(daemon): extract runner::lifecycle submodule (#2642)
Move run (main event loop) and tick (per-second scheduler) to
runner/lifecycle.rs. Clean up unused imports from mod.rs; Schedule
moved to cfg(test)-only alongside BuiltinTask and TaskAction.
